RentOwner Login and Logout now functional Although functional this still needs work such as UI and other use-case scenarios
There during this commit ther ewas a big concern for the developers not understandning the difference with render() and redirect() in views (From RentOwner login to home-page-view I demonstrated redirect then render since redirect can not carry context only render can).
For redirect then render approach to work I used built in django sessions as a payload used within django during redirect while render(context=) was used as data payload for templates.
Login and Logout was some makeshift code and was not using contrib.auth.authenticate(), this is however not a priority

Changes to be committed:
	modified:   GoRent_proj/database/models.py
	modified:   GoRent_proj/gildo/templates/gildo/search.html
	modified:   GoRent_proj/gildo/urls.py
	modified:   GoRent_proj/gildo/views.py
	modified:   GoRent_proj/registers/templates/registers/loginOwner.html
	modified:   GoRent_proj/registers/urls.py
	modified:   GoRent_proj/registers/views.py
	modified:   GoRent_proj/static/gildo/js/maps.js
	modified:   GoRent_proj/static/search/css/sidepanel_left.css
	new file:   GoRent_proj/static/search/css/sidepanel_right.css